Update Linear Kalman filter with a loop
As the real Hive tracker does, we have make a loop where position given by the Light House stand, and in the same time we observe derivation of  position given by the accelerometer. The display of variables doesn't work, but it's easy to draw graphs on Excel. The mean and the variance given by the algorithm allows to draw gaussians.